fix(ConnectedCornerstoneViewport): Fix definition of function map so it happens less frequently
`onMeasurementsChanged` currently defines a function map every time it is called. This function map
includes a throttle on the `modified` function which is not being used because the function is
called at most once per time it is instantiated. These changes elevate the map definition to the
file level, since what it is defining does not change based on the other executed code.