Add support for defining `whiteLabelling.logoComponent` as a function
This change is about white-labelling.
The template that generates the logo on the top of the page requires a
`React Element` and gets it from the `whiteLabelling.logoComponent` part of the global
configuration object.
Since `React` is not available in the scope of the global config object a React elmenet
can't be created on the configuration side.
These changes make it possible for the `whiteLabelling.logoComponent` configuration parameter
to be a **function** which is passed the React object.
This way it is possible to call React.createElement() on the configuration
side and replace the logo with a custom element.