fix(ohif-cli): Update generated files
Generated webpack, babel and jest configuration assume that extension or
mode are created in extensions directory which is not the case. This
commit updates the paths accordingly.
Base webpack requires to have a tailwind config in the directory, which
is not required. It is now used only if the file exists.