Adding references to two used markdown links
There could be more.  I would expect there might be an option in the documentation
system to error out if such links are not provided.  Also I would like to refer you
to a nice tool  https://linkcheck.github.io/linkchecker/  which could be used on CI
to once in a while check if references used in the documentation are still live.

Here is an example of use of it in BIDS specification project -- does "wonders" ;)
https://github.com/bids-standard/bids-specification/blob/HEAD/.circleci/config.yml#L52