refactor(config): Update AWS S3 Static wado server URLs
This commit updates the AWS S3 Static wado server URLs in the configuration files. The previous URLs were pointing to the cloudfront.net domain, and they have been updated to the dd14fa38qiwhyfd.cloudfront.net domain. This change ensures that the application can access the correct server for retrieving DICOM images.

Refactor the URLs in the following files:
- platform/app/public/config/customization.js
- platform/app/public/config/default.js
- platform/app/public/config/demo.js
- platform/app/public/config/e2e.js
- platform/app/public/config/kheops.js
- platform/app/public/config/local_static.js
- platform/app/public/config/multiple.js
- platform/app/public/config/netlify.js
- platform/app/public/config/public_dicomweb.js