feat(measurements): Provide for the Load (SR) measurements button to optionally clear existing measurements prior to loading the SR.
Refactored the load measurements code in OHIFCornerstoneSRMeasurementViewport.tsx to a new loadMeasurements command.
The SR measurements viewport status bar now extracts the Load button from the 'loadMeasurements' toolbar.
The Load button added to the toolbar defines the commands to be run. Typically this includes the loadMeasurements command
and optionally the clearMeasurements command.