BUG: Changing colormap for second volume not possible
When changing the colormap for the second volume, this always gets applied to the first volume. This is due to a bug in cornerstone extension commandsModule.ts, setViewportColormap. To set the colormap on a viewport, it requires the volumeId. Currently, this is retrieved by `viewport.getVolumeId()`, which only returns the first volumeId.

The event does pass the `displaySetInstanceUID`, which correctly identifies the volume. However, `volumeId` is slightly different (in this use case, prefixed by "CornerstoneStreamingVolume:". Implement a fix where all volumeIds in the viewport are filtered by comparing it to the displaySetInstanceUID. If a match is found (i.e. volumeId containing the displaySetInstanceUID), use that volumeId. Otherwise, use the first volumeId.