fix(segmentation): Adding segmentation via data portal in MPR mode.
Use volumeLoader.createAndCacheVolume to create the volume for a SEG display set.
This produces a StreamingImageVolume which is essential for switching to hanging protocols like MPR.
RTSTRUCT display sets are now marked as reconstructable.

For when multiple display sets occupy the active viewport, cache and restore each appropriately.

Added tests:
- launch MPR with unhydrated segmentation.
- launch MPR and then add (but not load) a segmentation from the overlay menu
- add (but not load) a segmentation from the overlay menu and then launch MPR