fix: 🐛 removed misleading isSupported from RTStruct metadata
There was a constant variable that was set to false and were intended to
be used as the flag for RTStructs, but they were always set to false and
were actually misleading. Also, the value itself wasn't being used
anywhere in the code. The isSupported from each individual countour is
kept since this information actually changes depending on the
ContourGeometricType