refactor: 💡 readability improvments at loadRTStruct function
There are some small improvments of readbility on this commit, such as
the change of a regular for loop to a for of. The usage of a Map instead
of a switch case to change a variable's name. Setting up the color array
as empty on the ROIContourData definition to have defined the full
structure of the object at definition.