feat(segmentation): Separate segmentation utilities from the segmentation tools.
Moved the segmentation utilities to the segmentation list section.

Styled the segmentation utility buttons as per design.

The segmentation utilities and tools for a specific representation type are now
enabled if there is a segmentation of the corresponding type loaded for the viewport.

Clicking a segmentation tool or utility button will activate the last selected
segmentation of the corresponding type.

The SegmentationService now keeps track of the last selected segmentation of a specific
representation type for a viewport.

Added the tab-contours icon.

Auto switch tabs whenever the first segmentation is added/loaded into a viewport.