fix: 🐛 change for loop due to unit test performance issues
This for loop was intentionally changed from a for...of loop to a
standard iterative loop, this because the old approach would create a
copy of the allRTStructData item, this takes more time and due to the
"fake async" nature of the function, causes a unit test to fail. The
standard iterative approach only creates a reference(pointer) to the
object, avoiding unecessary memory allocation.