fix: 🐛 removed unused logic at interleaveTopToBottom
This piece of code was removed due to the fact that it was not being
used anymore by the function after its latest changes, the changes
modified the way that the viewports are verified, thus removing the need
for this logic. Since it was not being used anymore, I opted for
removing it so it would not increase the complexity of a new reader
unecessarely.