fix: 🐛 reintroducing type parameter to addSegmentationRepresent
In a previous attempt to fix issue #5426 a change of how the
representationTypeToUse is acquired inside the
addSegmentationRepresention was made. The problem is that this change
removed the need for the usage of the type parameter, thus modifying the
API of the Segmentation Service that others can rely on. Not only that,
but implementing the fix like this would make RTStruct contours to error
out while loading.