fix: 🐛 adding viewport check to assing seg representation type
This change will ensure that the segmentation representation matches the
active viewport type. This is needed when you change from a Ortographic
viewport to a 3D one and vice versa. The idea is to get the segmentation
type and be sure it can be represented in the viewport, if not default
segmentation representation values will be assigned.

✅ Closes: #5426